move fetch_participant_and_session to utils
